Move logger to core module
